// packages/translator/src/codegen/hcl-writer.ts
export interface HclLiteral {
  readonly kind: 'literal';
  readonly value: string | number | boolean | null;
}

export interface HclExpr {
  readonly kind: 'expr';
  readonly expr: string;
}

export interface HclMapEntry {
  readonly key: string;
  readonly value: HclValue;
}

export interface HclMap {
  readonly kind: 'map';
  readonly entries: readonly HclMapEntry[];
}

export interface HclBlock {
  readonly kind: 'block';
  readonly body: readonly HclMapEntry[];
}

export interface HclList {
  readonly kind: 'list';
  readonly items: readonly HclValue[];
}

export interface HclBlocks {
  readonly kind: 'blocks';
  readonly blocks: readonly HclMapEntry[];
}

export type HclValue =
  | HclLiteral
  | HclExpr
  | HclMap
  | HclBlock
  | HclList
  | HclBlocks;